Changed acceleration, velocity, and position from integers to floats

+++ js/app.js
@@ -4,17 +4,17 @@ document.getElementById("plot").addEventListener("click", function() {
     let accel = document.getElementById("accel").value;
 
     let time = [0,1,2,3,4,5,6,7,8,9,10];
-    dist ? dist : dist= 0
-    iVel ? iVel : iVel = 0
-    accel ? accel : accel = 0
+    dist ? dist : dist= 0.0
+    iVel ? iVel : iVel = 0.0
+    accel ? accel : accel = 0.0
 
     pos = []
     for (let i = 0; i < time.length; i++) {
-        pos[i] = (iVel * time[i]) + ((0.5 * accel) * (time[i] ** 2))
+        pos[i] = (iVel * time[i]) + ((0.5 * accel) * (time[i] ** 2.0))
         console.log("(" + time[i] + ", " + pos[i] + ")")
     };
     console.log(dist)
-    let posi = pos.map(x => parseInt(x))
+    let posi = pos.map(x => parseFloat(x))
 
 
     let position = new Chart(document.getElementById("position"), {
@@ -33,7 +33,7 @@ document.getElementById("plot").addEventListener("click", function() {
 });
 velo = []
     for (let i = 0; i < time.length; i++) {
-    velo[i] = parseInt(iVel) + (parseInt(accel) * time[i])
+    velo[i] = parseFloat(iVel) + (parseFloat(accel) * time[i])
     }
     let velocity = new Chart(document.getElementById("velocity"), {
     type: 'line',
